+++ internal/box/cat/cat_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/slok/agebox/internal/model"
 	"github.com/slok/agebox/internal/secret/encrypt/encryptmock"
 	"github.com/slok/agebox/internal/secret/process/processmock"
+	"github.com/slok/agebox/internal/storage"
 	"github.com/slok/agebox/internal/storage/storagemock"
 )
 
@@ -53,7 +54,7 @@ func TestCatBox(t *testing.T) {
 			},
 			mock: func(m mocks) {
 				m.msp.On("ProcessID", mock.Anything, "secret1").Once().Return("secret1", nil)
-				m.mkr.On("GetPrivateKey", mock.Anything).Once().Return(nil, fmt.Errorf("something"))
+				m.mkr.On("ListPrivateKeys", mock.Anything).Once().Return(nil, fmt.Errorf("something"))
 			},
 			expErr: true,
 		},
@@ -64,7 +65,7 @@ func TestCatBox(t *testing.T) {
 			},
 			mock: func(m mocks) {
 				m.msp.On("ProcessID", mock.Anything, "secret1").Once().Return("secret1", nil)
-				m.mkr.On("GetPrivateKey", mock.Anything).Once().Return(nil, nil)
+				m.mkr.On("ListPrivateKeys", mock.Anything).Once().Return(&storage.PrivateKeyList{}, nil)
 
 				// Processed secret.
 				{
@@ -87,7 +88,7 @@ func TestCatBox(t *testing.T) {
 			},
 			mock: func(m mocks) {
 				m.msp.On("ProcessID", mock.Anything, "secret1").Once().Return("secret1", nil)
-				m.mkr.On("GetPrivateKey", mock.Anything).Once().Return(nil, nil)
+				m.mkr.On("ListPrivateKeys", mock.Anything).Once().Return(&storage.PrivateKeyList{}, nil)
 
 				// Processed secret.
 				{
@@ -108,7 +109,7 @@ func TestCatBox(t *testing.T) {
 			mock: func(m mocks) {
 				m.msp.On("ProcessID", mock.Anything, "secret1").Once().Return("secret1", nil)
 				m.msp.On("ProcessID", mock.Anything, "ignored").Once().Return("", nil)
-				m.mkr.On("GetPrivateKey", mock.Anything).Once().Return(nil, nil)
+				m.mkr.On("ListPrivateKeys", mock.Anything).Once().Return(&storage.PrivateKeyList{}, nil)
 
 				// Processed secret.
 				{
@@ -131,7 +132,7 @@ func TestCatBox(t *testing.T) {
 			},
 			mock: func(m mocks) {
 				m.msp.On("ProcessID", mock.Anything, "secret1").Once().Return("secret1", nil)
-				m.mkr.On("GetPrivateKey", mock.Anything).Once().Return(nil, nil)
+				m.mkr.On("ListPrivateKeys", mock.Anything).Once().Return(&storage.PrivateKeyList{}, nil)
 
 				// Processed secret.
 				{
@@ -147,7 +148,7 @@ func TestCatBox(t *testing.T) {
 			},
 			mock: func(m mocks) {
 				m.msp.On("ProcessID", mock.Anything, "secret1").Once().Return("secret1", nil)
-				m.mkr.On("GetPrivateKey", mock.Anything).Once().Return(nil, nil)
+				m.mkr.On("ListPrivateKeys", mock.Anything).Once().Return(&storage.PrivateKeyList{}, nil)
 
 				// Processed secret.
 				{
@@ -164,7 +165,7 @@ func TestCatBox(t *testing.T) {
 			},
 			mock: func(m mocks) {
 				m.msp.On("ProcessID", mock.Anything, "secret1").Once().Return("secret1", nil)
-				m.mkr.On("GetPrivateKey", mock.Anything).Once().Return(nil, nil)
+				m.mkr.On("ListPrivateKeys", mock.Anything).Once().Return(&storage.PrivateKeyList{}, nil)
 
 				// Processed secret.
 				{
@@ -181,7 +182,7 @@ func TestCatBox(t *testing.T) {
 			},
 			mock: func(m mocks) {
 				m.msp.On("ProcessID", mock.Anything, "secret1").Once().Return("secret1", nil)
-				m.mkr.On("GetPrivateKey", mock.Anything).Once().Return(nil, nil)
+				m.mkr.On("ListPrivateKeys", mock.Anything).Once().Return(&storage.PrivateKeyList{}, nil)
 
 				// Processed secret.
 				{
@@ -199,7 +200,7 @@ func TestCatBox(t *testing.T) {
 			},
 			mock: func(m mocks) {
 				m.msp.On("ProcessID", mock.Anything, "secret1").Once().Return("secret1", nil)
-				m.mkr.On("GetPrivateKey", mock.Anything).Once().Return(nil, nil)
+				m.mkr.On("ListPrivateKeys", mock.Anything).Once().Return(&storage.PrivateKeyList{}, nil)
 
 				// Processed secret.
 				{