+++ internal/secret/encrypt/encrypt.go
@@ -9,7 +9,7 @@ import (
 // Encrypter knows how to encrypt and decrypt secrets.
 type Encrypter interface {
 	Encrypt(ctx context.Context, secret model.Secret, keys []model.PublicKey) (*model.Secret, error)
-	Decrypt(ctx context.Context, secret model.Secret, key model.PrivateKey) (*model.Secret, error)
+	Decrypt(ctx context.Context, secret model.Secret, keys []model.PrivateKey) (*model.Secret, error)
 }
 
 //go:generate mockery --case underscore --output encryptmock --outpkg encryptmock --name Encrypter